Update code, tests and examples. Deleted a lot of old stuff.
