+++ examples/gems.locked
@@ -7,49 +7,63 @@ PATH
 GEM
   remote: https://rubygems.org/
   specs:
-    async (2.1.0)
-      console (~> 1.10)
-      io-event (~> 1.0.0)
-      timers (~> 4.1)
-    async-dns (1.3.0)
-      async-io (~> 1.15)
-    async-http (0.59.2)
-      async (>= 1.25)
-      async-io (>= 1.28)
-      async-pool (>= 0.2)
-      protocol-http (~> 0.23.1)
-      protocol-http1 (~> 0.14.0)
-      protocol-http2 (~> 0.14.0)
-      traces (>= 0.4.0)
-    async-io (1.34.0)
+    async (2.23.0)
+      console (~> 1.29)
+      fiber-annotation
+      io-event (~> 1.9)
+      metrics (~> 0.12)
+      traces (~> 0.15)
+    async-dns (1.4.1)
       async
-    async-pool (0.3.12)
+      io-endpoint
+    async-http (0.87.0)
+      async (>= 2.10.2)
+      async-pool (~> 0.9)
+      io-endpoint (~> 0.14)
+      io-stream (~> 0.6)
+      metrics (~> 0.12)
+      protocol-http (~> 0.49)
+      protocol-http1 (~> 0.30)
+      protocol-http2 (~> 0.22)
+      traces (~> 0.10)
+    async-pool (0.10.3)
       async (>= 1.25)
-    console (1.15.3)
-      fiber-local
-    fiber-local (1.0.0)
+    console (1.29.2)
+      fiber-annotation
+      fiber-local (~> 1.1)
+      json
+    fiber-annotation (0.2.0)
+    fiber-local (1.1.0)
+      fiber-storage
+    fiber-storage (1.0.0)
     geoip (1.6.4)
-    io-event (1.0.9)
-    nokogiri (1.13.8-arm64-darwin)
+    io-endpoint (0.15.1)
+    io-event (1.9.0)
+    io-stream (0.6.1)
+    json (2.10.1)
+    metrics (0.12.1)
+    nokogiri (1.18.2-arm64-darwin)
+      racc (~> 1.4)
+    nokogiri (1.18.2-x86_64-linux-gnu)
       racc (~> 1.4)
     process-daemon (1.0.1)
       rainbow (~> 2.0)
-    protocol-hpack (1.4.2)
-    protocol-http (0.23.12)
-    protocol-http1 (0.14.6)
+    protocol-hpack (1.5.1)
+    protocol-http (0.49.0)
+    protocol-http1 (0.30.0)
       protocol-http (~> 0.22)
-    protocol-http2 (0.14.2)
+    protocol-http2 (0.22.1)
       protocol-hpack (~> 1.4)
-      protocol-http (~> 0.18)
-    racc (1.6.0)
+      protocol-http (~> 0.47)
+    racc (1.8.1)
     rainbow (2.2.2)
       rake
-    rake (13.0.6)
-    timers (4.3.4)
-    traces (0.7.0)
+    rake (13.2.1)
+    traces (0.15.2)
 
 PLATFORMS
   arm64-darwin-21
+  x86_64-linux
 
 DEPENDENCIES
   async-http
@@ -59,4 +73,4 @@ DEPENDENCIES
   rubydns!
 
 BUNDLED WITH
-   2.3.10
+   2.6.2
