--- examples/gems.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-# Released under the MIT License.
-# Copyright, 2014-2022, by Samuel Williams.
-# Copyright, 2014, by Peter M. Goldstein.
-
-source "https://rubygems.org"
-
-gem "rubydns", path: "../"
-gem "process-daemon"
-gem "nokogiri"
-
-gem "async-http"
-
-gem "geoip"
