+++ lib/rubydns.rb
@@ -14,8 +14,14 @@ module RubyDNS
 	Resolver = Async::DNS::Resolver
 	
 	# Run a server with the given rules.
-	def self.run (*arguments, server_class: Server, **options, &block)
-		server_class.new(*arguments, **options, &block).run
+	def self.run(*arguments, server_class: Server, **options, &block)
+		server = server_class.new(*arguments, **options)
+		
+		if block_given?
+			server.instance_eval(&block)
+		end
+		
+		return server.run
 	end
 	
 	# @deprecated Use {RubyDNS.run} instead.
