+++ readme.md
@@ -32,25 +32,21 @@ Here is the code from `examples/basic-dns.rb`:
 #!/usr/bin/env ruby
 require 'rubydns'
 
-INTERFACES = [
-	[:udp, "0.0.0.0", 5300],
-	[:tcp, "0.0.0.0", 5300],
-]
+# Use the system default resolver for upstream queries:
+upstream = Async::DNS::Resolver.default
 
-IN = Resolv::DNS::Resource::IN
+# We will use port 5300 so we don't need to run the server as root:
+endpoint = Async::DNS::Endpoint.for("localhost", port: 5300)
 
-# Use upstream DNS for name resolution.
-UPSTREAM = RubyDNS::Resolver.new([[:udp, "8.8.8.8", 53], [:tcp, "8.8.8.8", 53]])
-
-# Start the RubyDNS server
-RubyDNS::run_server(INTERFACES) do
-	match(%r{test.local}, IN::A) do |transaction|
+# Start the RubyDNS server:
+RubyDNS.run(endpoint) do
+	match(%r{test.local}, Resolv::DNS::Resource::IN::A) do |transaction|
 		transaction.respond!("10.0.0.80")
 	end
 
 	# Default DNS handler
 	otherwise do |transaction|
-		transaction.passthrough!(UPSTREAM)
+		transaction.passthrough!(upstream)
 	end
 end
 ```
